requirements.txt
flask==3.0.3
langchain==0.2.1
google-generativeai==0.3.2
langchain-community==0.2.1
pymysql==1.1.1
langchain_google_genai==0.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Security==3.0.0
gunicorn==21.2.0
redis==5.0.1
bcrypt==4.1.2
python-jose==3.3.0
pytest==8.0.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pytest-cov==4.1.0